// frontend/src/app/generated/snippets-api.ts
export interface components {
    schemas: {
        ApiError: {
            code: string;
            fields: {
                [key: string]: string;
            };
            message: string;
        };
        CreateRequest: {
            content: string;
            /** @enum {string} */
            language: "JSON" | "TEXT";
            title: string;
        };
        Page: {
            hasNext: boolean;
            items: components["schemas"]["Summary"][];
            /** Format: int32 */
            limit: number;
            /** Format: int32 */
            offset: number;
        };
        Snippet: {
            content: string;
            /** Format: date-time */
            createdAt: string;
            /** Format: uuid */
            id: string;
            /** @enum {string} */
            language: "JSON" | "TEXT";
            title: string;
            /** Format: date-time */
            updatedAt: string;
            /** Format: int64 */
            version: number;
        };
        Summary: {
            /** Format: uuid */
            id: string;
            /** @enum {string} */
            language: "JSON" | "TEXT";
            title: string;
            /** Format: date-time */
            updatedAt: string;
            /** Format: int64 */
            version: number;
        };
        UpdateRequest: {
            content: string;
            /** @enum {string} */
            language: "JSON" | "TEXT";
            title: string;
            /** Format: int64 */
            version: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}